move stepper package location to outside cloud directory
